+++ MajorPrivacy/Views/AccessView.cpp
@@ -19,7 +19,7 @@ CAccessView::CAccessView(QWidget *parent)
 	//connect(m_pTreeView, SIGNAL(ResetColumns()), this, SLOT(OnResetColumns()));
 	//connect(m_pTreeView, SIGNAL(ColumnChanged(int, bool)), this, SLOT(OnColumnsChanged()));
 
-	connect(m_pItemModel, SIGNAL(CheckChanged(const QModelIndex&, bool)), this, SLOT(OnCheckChanged(const QModelIndex&, bool)));
+	//connect(m_pItemModel, SIGNAL(CheckChanged(const QModelIndex&, bool)), this, SLOT(OnCheckChanged(const QModelIndex&, bool)));
 
 	QByteArray Columns = theConf->GetBlob("MainWindow/AccessView_Columns");
 	if (Columns.isEmpty()) {