+++ MajorPrivacy/Views/ExecutionView.cpp
@@ -17,7 +17,7 @@ CExecutionView::CExecutionView(QWidget *parent)
 	//connect(m_pTreeView, SIGNAL(ResetColumns()), this, SLOT(OnResetColumns()));
 	//connect(m_pTreeView, SIGNAL(ColumnChanged(int, bool)), this, SLOT(OnColumnsChanged()));
 
-	connect(m_pItemModel, SIGNAL(CheckChanged(const QModelIndex&, bool)), this, SLOT(OnCheckChanged(const QModelIndex&, bool)));
+	//connect(m_pItemModel, SIGNAL(CheckChanged(const QModelIndex&, bool)), this, SLOT(OnCheckChanged(const QModelIndex&, bool)));
 
 	QByteArray Columns = theConf->GetBlob("MainWindow/ExecutionView_Columns");
 	if (Columns.isEmpty()) {